+++ backend/settings/settings_test.go
@@ -42,7 +42,6 @@ func TestConfigLoadSpecificValues(t *testing.T) {
 	}{
 		{"Auth.Method", Config.Auth.Method, newConfig.Auth.Method},
 		{"Auth.Method", Config.Auth.Method, newConfig.Auth.Method},
-		{"UserDefaults.HideDotfiles", Config.UserDefaults.HideDotfiles, newConfig.UserDefaults.HideDotfiles},
 		{"Server.Database", Config.Server.Database, newConfig.Server.Database},
 	}